+++ src/components/AssignedUsers/constants.js
@@ -1,15 +1,15 @@
 import { FormattedMessage } from 'react-intl';
 
-export const VISIBLE_COLUMNS = ['fullName', 'patronGroup'];
+export const VISIBLE_COLUMNS = ['fullName', 'groupName'];
 
 export const COLUMN_MAPPING = {
   fullName: <FormattedMessage id="ui-users.permissions.assignedUsers.name" />,
-  patronGroup: <FormattedMessage id="ui-users.permissions.assignedUsers.patronGroup" />,
+  groupName: <FormattedMessage id="ui-users.permissions.assignedUsers.patronGroup" />,
 };
 
 export const COLUMN_WIDTH = {
   fullName: '50%',
-  patronGroup: '50%',
+  groupName: '50%',
 };
 
 export const USERS_API = 'users';